requirements.txt
# --- Backend Core ---
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1

# --- AI ---
ag2[openai]==0.3.1  # AutoGen framework with OpenAI-compatible client

# --- Frontend ---
streamlit==1.32.2
requests==2.31.0